requirements.txt
blinker==1.9.0
click==8.3.1
etcd3==0.12.0
Flask==3.1.2
grpcio==1.76.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
protobuf==3.20.3
python-dotenv==1.2.1
six==1.17.0
tenacity==9.1.2
typing_extensions==4.15.0
Werkzeug==3.1.4
